Add $Junk/$NotJunk/$Phishing keyword support (RFC 9051 §2.3.2)
Parse, store, and send the three standard junk classification
keywords.  Widen SBitFlags from unsigned long to uint64_t so enum
bits 32-34 are correct on all platforms (1L << 32 is undefined on
Win32 LLP64).  Offline cache and action recording serialization
widened to persist 64-bit flags; version gate at 0x0C handles
upgrade from pre-0x0C caches.

Enforce mutual exclusivity: when both $Junk and $NotJunk are
present on a message, treat as neither set (MUST per RFC 9051).

Also fix copy_uids not cleared on TRYCREATE retry in _CopyMessage,
and remove redundant catch(...) in _AppendMbox.

+++ Sources_Common/Offline/Mail/CLocalClient.h
@@ -149,11 +149,11 @@ protected:
 
 	struct SIndexRecord
 	{
-		// All members must be uint32_t for correct serialization on 64-bit
-		// (sizeof(unsigned long) = 8 on LP64, but file format is 4 bytes per field)
+		// All members are uint32_t for correct disk serialization, except mFlags
+		// which is uint64_t (serialized as two uint32_t values, lo then hi)
 		uint32_t mCache;
 		uint32_t mIndex;
-		uint32_t mFlags;
+		uint64_t mFlags;
 		uint32_t mUID;
 		uint32_t mLocalUID;
 		uint32_t mMessageStart;
@@ -181,9 +181,9 @@ protected:
 			{ return mIndex; }
 		const uint32_t& Index() const
 			{ return mIndex; }
-		uint32_t& Flags()
+		uint64_t& Flags()
 			{ return mFlags; }
-		const uint32_t& Flags() const
+		const uint64_t& Flags() const
 			{ return mFlags; }
 		uint32_t& UID()
 			{ return mUID; }
